rename(#6): 버튼 스타일 이름 변경
